refactor: update interfaces
